+++ src/phosphorus/lib/types.py
@@ -6,6 +6,7 @@ if TYPE_CHECKING:
     from phosphorus.lib.packages import Package
     from phosphorus.lib.requirements import ResolvedRequirement
     from phosphorus.lib.resolver import LockEntry
+    from phosphorus.lib.versions import Version
 
 Match = Optional[str]  # TODO (py3.9): Use |
 
@@ -19,3 +20,9 @@ class ResolvedPackageInfo(TypedDict):
 class PackageDiff(TypedDict):
     update: set[tuple[ResolvedRequirement | None, LockEntry]]
     remove: set[ResolvedRequirement]
+
+
+class InstalledVersions(TypedDict, total=False):
+    package: Package
+    installed_version: Version
+    latest_version: Version